pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "py-logger-alert" # Escolha um nome único no PyPI
version = "0.1.5"
authors = [
  { name="Vitor Marinheiro", email="vitormarinheiro9@gmail.com" },
]
description = "Um monitor de logs automático via Telegram para robôs Python"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "requests",
    "python-dotenv",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["py_logger_alert*"]